Show siblings in navigation when currentFolderOnly is not set.
When the option "Only show the contents of the current folder." in the navigation
portlet configuration is not activated, show also siblings in the navigation.

This change means that existing navigation portlets might need to be reconfigured
when the siblings should not be shown, since the default value of this option is
to show siblings.