Remove unused sass-watcher stuff. The css will be generated automatically.
Current sass watcher is not working for building statis scss files.
According to 82fdefd the scss files are generated through the server.
So the sass watcher script is no longer needed. btw viewing the index.html file without running a server
does not work anyway so the statis files are obsolete.