Provide a base tag.
Plone has removed the base tag. The base tag is used by the browser
for generating URLs based on relative paths.
In Plone it is possible to get the same resource with and without a
trailing slash.
When a page contains relative paths (e.g. HTML block) the behavior is
incosistent when the page is loaded with or without a trailing slash.

The fix is to simply add the base tag again.
Plone 4 used to check whether the content "isPrincipiaFolderish", but
since all our content matches this criteria we are always adding a
trailing slash.