Fix position of the navbar items
Some elements are wrapped in a div which doesn't have the same
font-size as the elements in it. And because the margin is set in em it is
relative to the font size. This resulted in a 1px difference between the
margins causing it to not be aligned in some zoom levels.